Build/run mobile canvas_test only for mobile build tag
This requires moving TestMain and mobileApp to a different test file.
