+++ internal/driver/mobile/driver_test.go
@@ -2,15 +2,44 @@ package mobile
 
 import (
 	"image/color"
+	"os"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/assert"
 
 	"fyne.io/fyne/v2"
 	fynecanvas "fyne.io/fyne/v2/canvas"
 	"fyne.io/fyne/v2/container"
+	"fyne.io/fyne/v2/internal/async"
 	"fyne.io/fyne/v2/widget"
 )
+
+var d *driver
+
+func TestMain(m *testing.M) {
+	currentApp := fyne.CurrentApp()
+	tester := newTestMobileApp()
+	d = tester.Driver().(*driver)
+	d.queuedFuncs = async.NewUnboundedChan[func()]()
+	fyne.SetCurrentApp(tester)
+
+	waitForStart := make(chan struct{})
+	go func() {
+		// Wait for app loop to be running (plus a moment in case of scheduling switches).
+		<-waitForStart
+
+		// Just like the GLFW tests, wait a short while for the driver to start
+		time.Sleep(time.Millisecond * 100)
+
+		ret := m.Run()
+		fyne.SetCurrentApp(currentApp)
+		os.Exit(ret)
+	}()
+
+	close(waitForStart) // Signal that execution can continue.
+	tester.Run()
+}
 
 func Test_mobileDriver_AbsolutePositionForObject(t *testing.T) {
 	for name, tt := range map[string]struct {
@@ -78,5 +107,31 @@ func Test_mobileDriver_AbsolutePositionForObject(t *testing.T) {
 			got := d.AbsolutePositionForObject(o)
 			assert.Equal(t, tt.want, got)
 		})
+	}
+}
+
+type mobileApp struct {
+	fyne.App
+	driver fyne.Driver
+}
+
+func (a *mobileApp) Driver() fyne.Driver {
+	return a.driver
+}
+
+func (a *mobileApp) Run() {
+	// This is an incomplete driver loop - our CI does not currently support booting the mobile graphics
+	// TODO replace with a full mobileApp.Run() once that is resolved
+	async.SetMainGoroutine()
+
+	for fn := range d.queuedFuncs.Out() {
+		fn()
+	}
+}
+
+func newTestMobileApp() fyne.App {
+	return &mobileApp{
+		App:    fyne.CurrentApp(),
+		driver: NewGoMobileDriver(),
 	}
 }
